TEIID-714 TEIID-687 TEIID-685 TEIID-684 adding the ability to import procedures and fixing the DatabaseMetadata.getProcedures call.

+++ metadata/src/main/java/org/teiid/metadata/index/RecordFactory.java
@@ -806,7 +806,7 @@ public class RecordFactory {
         paramRd.setPosition(Integer.parseInt((String)tokens.get(tokenIndex++)) );
 
         // The next token is parameter type        
-        paramRd.setType(Integer.parseInt((String)tokens.get(tokenIndex++)));
+        paramRd.setType(Short.parseShort((String)tokens.get(tokenIndex++)));
 
         // The next token is flag for parameter optional prop
         char[] flags = ((String)tokens.get(tokenIndex++)).toCharArray();
